*** empty log message ***

+++ coders/pnm.c
@@ -299,7 +299,7 @@ static Image *ReadPNMImage(const ImageInfo *image_info,ExceptionInfo *exception)
         CloseBlob(image);
         return(image);
       }
-    if ((image->columns*image->rows) == 0)
+    if ((size_t) (image->columns*image->rows) == 0)
       ThrowReaderException(CorruptImageWarning,
         "Unable to read image: image dimensions are zero",image);
     scale=(Quantum *) NULL;
